allow custom output dir and dpi for plots
